Fix bug with alternative_format_provider field
This logic is faulty. With the previous logic an option is selected in
two circumstances.

1. If the alternative_format_provider_id on the edition matches the
value of the option
2. If the users organisation's id matches the options id

Both of these can be correct, but not concurrently.

At the moment, as two options are being marked as selected, the
organisation that falls later in the alphabet is being marked as selected.

The correct behaviour is:

When an alternative_format_provider_id is present on an edition, we
should use that to select the correct option.

When alternative_format_provider is nil, we should select the users
organisation.