Change validation rules for reslugging documents
At the moment, we prevent a document from having it's slug changed if
the new slug exists of any document. However Whitehall permits
duplicated slugs, provided the document type is different. This is
confirmed by the uniqueness in the schema [1] and the rules that are
checked when a document is first created.

Therefore updating the reslugging rules to permit a duplicate slug
provided the document type is different.

1: https://github.com/alphagov/whitehall/blob/9ea7dae5bfc445eaa56c682b244dc25bc4c5827d/db/schema.rb#L199
2: https://github.com/alphagov/whitehall/blob/9ea7dae5bfc445eaa56c682b244dc25bc4c5827d/app/models/document.rb#L59-L73