Dynamically show and hide fields based on News Article Type
When a user selects the WorldNewsStory subtype from the type
select, certain fields need to be hidden and cleared and others need
to be shown. This adds in that functionality.

When the WorldNewsStory type is selected the following behaviour
should happen:

1. the world news location select should show
2. the document language fields should show
3. the ministers select should be hidden and the value should be reset
4. the org fields should be hidden and all selects' values should be
reset

When another type is selected (or nil) the inverse should happen

1. the the world news location select should be hidden and reset
2. the document language fields should be hidden and reset (this
has already been implemented)
3. the minsters select should show
4. the org fields should show

Co-authored-by: Conor Lowney <conor.lowney@digital.cabinet-office.gov.uk>