Remove locales from presenter for ministers page
The page is only presented and rendered in English, so we don't need to
include paths for other locales or test for a translated content item.

The reason for this is we don't have translations for the names of most
roles, therefore the page is only currently rendered in English.