Add rake task for cloning a HTML Attachment from published to draft edition
We've run into an issue a couple of times now where a user has created
a new draft edition and the HTML Attachment has not been carried across
due to the Govspeak being invalid.

This can occur when updates happen to the Govspeak gem between the HTML
Attachment being created and a new edition draft edition being created
as the validator only validates the govspeak content when a change is
made.

While we can fix the govspeak then ask the user to delete and re-create
the draft edition, if they've started to make changes it could cause
them a bit of hassle. This rake task hopefully alleviates that issue
by building out the functionality to clone the attachment across
from a published edition to a draft edition.

I've personally had this happen to me twice so hopefully this will
make it easier to fix going forward.