Remove ministers index page from links in roles
At the moment, we are including a link to the ministers index page in
the links for every ministerial role (including past roles).

This results in a reverse link in the ministers index page content item
that has become too large to work with.

In later work, this will be replaced with a link to the current and
relevant ministers only in the index page's content item.

We are having to make this change in two steps (step one: empty the
array, step two: remove the `ministerial` link completely), as patching
the links won't remove the link if the key is not included. A later PR
will remove the key completely from the presenter.