Create Publishing API JSON for embassies index page
We've recently added a new document type `embassies_index` to the
Publishing API[1]. This commit ensures we can generate a JSON document
that validates against the publishing schema for that document type
from the data in Whitehall.

We're setting `rendering_app` to "whitehall" so that we can publish
the `embassies_index` document to the content store first and then
subsequently update Collections to render that content. When that has
happened, we can set the `rendering_app` to collections in this class
to complete the migration.

A few notes to explain some of the choices in this commit:

- we need to include `update_type` even though it's not validated by the
schema. It's validated in the code of publishing-api instead[2].
- we've decided to explicitly set the locale via `BaseItemPresenter`'s
defaults even though this isn't required by the schema.
- we're not setting `public_updated_at` or `description` because these
properties are optional[3] and we don't yet know if the rendering app
will need them.
- there's already a "static page" with a content ID for /world/embassies
created by lib/publish_static_pages.rb. We've reused that ID in this
presenter (and will remove the static page publisher in a subsequent commit)

[1] https://github.com/alphagov/publishing-api/pull/2354
[2] https://github.com/alphagov/publishing-api/blob/4f0070b09cdafbf8f10e727481e1edc097c1cd3c/app/commands/v2/publish.rb#L92
[3] https://github.com/alphagov/publishing-api/blob/main/docs/api.md#json-attributes

Co-authored-by: Chris Lowis <chris.lowis@digital.cabinet-office.gov.uk>