Rename methods for clarity
The /world/embassies page is a collection of links to where British
nationals can best find assistance. There's nothing in this model
which cares about the specific services that an embassy provides (for
example, we're not looking at the WorldwideOfficeWorldwideServices that
a WorldwideOffice has). We think this set of method renames makes it
clearer what the intention of these methods is.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>