Remove assertion using Embassy#offices
The offices method is currently only used by the Embassy class and
these tests. It's only used in Embassy to check whether there are
offices that exist or not - we do not use the offices from the array
elsewhere in the code. We're confident that nothing is
relying on the title of the office and that the current test makes
sure things are behaving as they should. We're removing this
assertion that uses offices to allow us to make offices private and
simplify the interface of Embassy in a subsequent commit.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>