Add tests for Embassy#remote_office
We think these new tests explain more clearly what is being tested and
rely on a single interface (#remote_office) into the underlying
implementation, so we're also removing the explicit tests for
`remote_office_country` so we can subsequently make it private (the
tests are the only place that calls this method outside of the Embassy
class).

We've decided to inline the setup into each test for now to make it
clearer to us where setup is shared and where it differs. We hope to
factor this out as appropriate later as we continue to increase the
test coverage.