Rewrite tests for Embassy class
We use contexts to describe how instances of the Embassy class respond
given the various states it can be initialized in based on data we've
encountered in production. These tests are almost certainly not
comprehensive but they give us quite a lot of confidence that things are
working as they should.

Co-authored-by: Chris Lowis <chris.lowis@digital.cabinet-office.gov.uk>