Fix a bug in Embassy#remote_services_country
We noticed it was possible for this method to return the incorrect
country if there was a non-embassy office in a third country. We
haven't seen this bug cause any issues, but it makes the
remote_services_country method easier to understand if we fix it.

We noted this problem in the commit titled "Extract
Embassy#offices_in_location?". Subsequent refactorings and tests have
finally given us the confidence to make the change.

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>