Move `AuditTrail` out of the `Edition` namespace
We want to be able to create audit trails for non-editionable models.

There is an existing audit trail implementation that is only used for
editionable models. However, as it is quite generic we want to repurpose it so
that it can be used for both non-editonable and editionable models.

This moves the `AuditTrail` concern out of the `Edition` namespace.

The `alert!` method of the `AuditTrail` is possibly non-generic, this will be
refactored or removed  at a later point.