Ensure featured link fields are present on world_location_news edit page
At the moment, if a user arrives on the edit page with no
featured links they are unable to add any. As none are present, the
fields don't get created in the partial.

This build a blank featured link on the world_location_news if none
are present before rendering the edit view.