Update ErrorSummaryComponent to accept errors objects
Currently, we have an issue with duplicate error messages being
present on a WorldLocationNews object on the edit world location news
page due to both associations accepting nested attributes for each other.

This updates the the ErrorSummaryComponent to be able to take 3 types
of objects

1. a model
2. a models errors (model.errors)
3. an array of errors. This is needed as we need to be abe to filter out
duplicate errors using `reject` which returns an array of objects.