Show an audit trail for worldwide organisations
Now that we create an audit trail for worldwide organisations, we want to
display the history on the admin page.

ViewComponents for editionable models already exist. But, because these contain
edition specific logic, rather than modifying them, we want to create separate
ViewComponents for non-editionable models .

This adds a paginated audit trail alongside the details of the worldwide
organisations. Any initial audits are shown as "No history before this time" by
"User (unknown)"