Remove presence validation for alt text from feature
Alt text isn't required for recommended for decorative images.
This removes the validation that enforces presence for it in the
features model.

It also ensures that an empty string is  passed for the alt text in the
FeaturedDocumentsPresenter so that an empty alt text attribute is
rendered to the user to show that the image in described in the
surrounding text.