Create initial audits for worldwide organisations
We will soon be auditing worldwide organisations; this will create a `Version`
record when a worldwide organisation is created or updated. However, until
auditing is added, we have no way of knowing what changes were made or by
whom.

This migration creates a `Version` with an "initial" event type for all worldwide
organisations. We can then handle this in the UI in an appropriate way by
displaying "No history before this time".

We want to deploy the worldwide organisation auditing work in two steps -
first, by running this migration, then, by merging separate work that adds audits.
There may be audits that are not recorded in the period between the two steps.
We have decided that this is only a minor issue, and we will attempt to deploy
the follow up work shortly after this migration is run in order to reduce the
likelihood of this happening.