Make image validation errors more meaningful
Update the image validator so it differentiates between images which are
too small and those which are too large.

The new :too_small and :too_large error types allow us to
programatically detect the reason for an uploading image being
considered invalid.

This enables us to perform different actions based on the reason for the
validation error - e.g. the controller can redirect to a 'crop' page if
the image is too large, but still show the error message if it's too
small.

The error messages rendered to users have also been tweaked slightly to
make them more readable. Instead of just listing the uploaded image's
dimensions, they now clearly state if the image is "too small" or "too
large".