Fix 'flash of unstyled content' on subnav menus
This change fixes a brief [flash of unstyled content][1] that caused an
unsightly flickering of the subnav menus on page load.

The no-JS experience is for all subnav menus to be expanded and visible.
For JS-enabled users, the subnavs will hide immediately once the
JavaScript module 'navbar-toggle' has initialised.

However there's an 'in between' period where the HTML document is still
loading, so the JS modules haven't initialised yet, and the subnav menus
are visible on screen.

To handle this case, I've introduced the `.hide-before-js-module-init`
class which will be hidden so long as JavaScript is enabled. JS modules
can then remove the class once they're styled elements as required.

[1]: https://en.wikipedia.org/wiki/Flash_of_unstyled_content