Add Currently featured tab in the GOV.UK Design System
This adds a CurrentlyFeaturedTabComponent View Component which renders
and tests the currently featured tab on the world_location_news
features page.

I've chosen to use a View Component as there's quite a decent
amount of complexity going on.

A feature can belong to a document, topical event or offsite link. Due,
to this we have to render the content for each row conditionally, based
on the association. It feels like something worth unit testing.

It follows the design laid out in the design card
https://trello.com/c/mgckMCcS/809-currently-featured-design-placeholder