Add unit test for PublishingApi::PayloadBuilder::Routes
The additional_routes logic was added in this commit [1], but its effect
was only tested in the PublishingApi::OrganisationPresenter test. I
think it's useful to include a more specific unit test for this Routes
class, particularly because I'm about to make changes to it.

I've also tweaked the name of the original test method to make its
purpose clearer and to differentiate it from the new test method.

[1]: 381856144a7797277f9690184685bffd698019af