Add basic implementation of the documents tab on the features page
This adds the documents tab to the world news story features page.
It ensures the per page of the results is 15 to be in line with
the documents index page.

There's some issues with the anchoring of the page to the tab
on page load when filtered or the pagination links are clicked.
These will be handled in a follow up PR.

I've removed the topical events tab as world_location_news objects
cannot have topical events for features.