Allow paginator to accept an anchor tag and update PaginationHelper
We need to be able to pass the paginator an anchor so that when a
pagination link is clicked it opens on the documents tab.

This PR updates the govuk_paginator to accept an optional anchor
paramater.

It also updates the PaginationHelper to handle urls with anchors being
passed in.

At the moment, when a page query string paramater is not present on the
url passed in, the pagination helper constructs one and appends
it to the url.

If an anchor is assed in this doesn't work as it does something like

/edition/12345?random_query_string=random#tab_id&page=2

This updates the helper insert the query string parameter before the
anchor.