Fix bug in edition filter where per_page is not passed to Kaminari
At the moment, we're not passing through the per page argument to
Kaminari so it's defaulting to 25 per page when it builds the
pagination.

This means that if there if there are 20 documents, 15 render on
the page and there's no way to access the other 5 as it thinks
there's only 1 page of documents.