Remove unused params from permitted WO translation params
The summary, description & services fields were removed from the
corresponding form in this commit [1] as part of this PR [2].

Currently WorldwideOrganisation#summary is derived from its associated
About page via HasCorporateInformationPages and not from a translated
attribute on WorldwideOrganisation itself. It looks as if the summary
was moved onto the About page in that same PR [2].

WorldwideOrganisation#description & #services were concatenated into the
body of the About page in this commit [3] on the same PR [2].

While I doubt this is causing any warnings or errors, it makes the code
unnecessarily confusing, particularly as we currently planning to move
the summary & body back from the About page onto the
WorldwideOrganisation itself! See this Trello card [4].

[1]: 1be1339088f6dd7fdd814ac926cb7abfe2334741
[2]: https://github.com/alphagov/whitehall/pull/1476
[3]: b12e7bf5cbc6161e9cb28810373c204767be2a54
[4]: https://trello.com/c/hwo2GEzk