Remove redundant worldwide_organisation_translations.services
WorldwideOrganisation#services were concatenated into the body of the
About page in this commit [1] as part of this PR [2] and so this
database column is no longer needed. There don't appear to be any
references to it left in the code.

In particular I've carefully checked the following templates:

* app/views/admin/worldwide_organisations/show.html.erb
* app/views/admin/worldwide_organisations_translations/edit.html.erb
* app/views/worldwide_organisations/show.html.erb

And these presenters:

* WorldwideOrganisationPresenter
* Api::WorldwideOrganisationPresenter
* PublishingApi::WorldwideOrganisationPresenter

[1]: b12e7bf5cbc6161e9cb28810373c204767be2a54
[2]: https://github.com/alphagov/whitehall/pull/1476