Capture stdout & stderr in worker tests
By swallowing this expected output we can reduce the noise in the test
output, which should mean we're more likely to see any *unexpected*
output.

We're already using MiniTest::Assertions#assert_output, so it doesn't
seem to controversial to use MiniTest::Assertions#capture_subprocess_io
[1] which is already available in all these tests.

I've kept the scope of the capture_subprocess_io block as small as
possible to reduce the chance that we miss some important output.

I think it's safe to assume we don't care about the output from these
tests since it has been appearing for some considerable time in every
build, but if we do care we should make explicit assertions about the
output we're expecting.

[1]: https://docs.seattlerb.org/minitest/Minitest/Assertions.html#method-i-capture_subprocess_io