Use progress format for pact:verify output
By default the pact:verify rake task provided by the pact gem uses the
(very verbose) RSpec "documentation" format. This seems at odds with the
"progress"-style format that we're using in the main test suite and I
think it clutters up the test output unnecessarily.

Irritatingly I couldn't find a way to override the formatter used by the
pact:verify task, so I've had to create a new one which shells out to
the pact CLI. I've wired this new task into the default task, but the
original pact:verify task is still available in case someone wants to
run it.