Avoid RestClient warning from RummageableTest
The following warning was being generated by the "amend should post
amendments to a document by its link" test in RummageableTest:

    warning: Overriding "Content-Type" header "application/json"
    with "application/x-www-form-urlencoded" due to payload

I'm not 100% sure, but I think this was happening because the `changes`
the test is passing to `Whitehall::Rummageable::Index#amend` are a Hash;
whereas `RestClient.send` is expecting the payload of the request to be
a String.

Other methods like `Index#add` & `Index#commit` encode the payload using
`MultiJson.encode` and so it seems reasonable that `Index#amend` does
that same. What's more having made this change, the expected call to
`Net::HTTP` as specified by `assert_requested` in the test is still
being made and so the test still passes.

It's slightly perplexing that this hasn't been identified before - I can
see that this apparent discrepancy between `Index#add` & `Index#amend`
seems to be present even in the last version of the rummageable gem [1]
before it was copied into whitehall [2]. However, as far as I can see
the only call to `Index#amend` in the whitehall code is this call from
RummageableTest, so perhaps we could remove it altogether...?

[1]: https://github.com/alphagov/rummageable/blob/225fe03bba5528dc4495b5a1be63396438713229/lib/rummageable.rb#L35
[2]: https://github.com/alphagov/whitehall/pull/2960