Add statistics announcements publications index page to support move to GOV.UK Design System
—— 

* Added statistics_announcement_publications_controller and test
* Added statistics_announcement_publications to routes
* Added statistics_announcment_publications index page  

It is possible to connect documents to statistics announcements. To date,
this functinality has been implemented with the "GOVUK.documentFinder"
javascript.

We are moving away from this, to a statically rendered search, which lists
the relevant documents.

This commit implements the first step in the new flow, rendering a search bar.

https://trello.com/c/HNwwVfU0