Add access and opening time columns to worldwide organisation and office
We want to remove the `access_and_opening_times` table to make it easier to
move the rendering of Worldwide Organisations and Worldwide Offices out of
Whitehall.

Create migrations to add new columns to `worldwide_organisations` and
`worldwide_offices`. Work will follow to use these instead of the
`access_and_opening_times` table.