Clarify tag on a Bootstrap feature test
Switch the tag from @design-system-wip to @bootstrap-only.

While those tags are functionally the same, @bootstrap-only signifies
that a test is deliberately excluded from the Design System feature test
run because it does not apply.

In this case, images are uploaded via the new "Images" tab rather than
directly on the Edition "edit" page. Therefore this test isn't needed,
because the JavaScript "add another image" functionality no longer
exists. This only applies to Bootstrap.