Remove embassies index static page from PublishStaticPages
As of the previous commit we now have "real" JSON to send to the
Publishing API for /world/embassies, so we no longer need this "static
page" definition.

After discussion with @BeckaL we believe that these static pages were
part of some prior work to migrate rendering from Whitehall to other
apps that was never completed, so we don't believe there's any
unintended consequences from removing this static page definition.