Add hooks to republish embassies_index when data changes
This commit ensures that the `embassies_index` document is republished
to the Publishing API whenever one of the models that it depends on
for its data is created, updated or destroyed.

We've decided not to do any clever change detection in these hooks to
only republish the document if a specific piece of data we depend
on (for example the name of the WorldwideOrganisation) changes. We
think the underlying data changes rarely enough that the added
complexity to reduce the number of API requests isn't justified.

The creation of WorldwideOffice models in the factory definitions causes
a dependent Contact model to be updated, which is why we've had to
those API requests twice. We suspect it's also possible in production that 2
API requests could be made when a new WorldwideOffice is created but
we haven't done anything to mitigate a potential race condition with
the arrival of those requests at the Publishing API. We don't think
it'll cause an issue, but leaving a note here in case we're wrong :)

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>