Instantiate RemoteOffice with a string for location
Before this commit `location` was an object that casts to the
country's name when rendered by the template. With this commit we
explicity use the `name` instead. We want to move the rendering of the
page that uses this model (/world/embassies) out of Whitehall and into
Collection, and pass the data we require to do so via the Publishing
API. It makes subsequent JSON type validation easier if we're explicit
about the values.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>