Migrate Access and Opening Times data
We have added new columns to Worldwide Organisations and Worldwide Offices so
that we can remove the access and opening times table.

Copy the data over to the new columns. This specifically uses `update_columns`
instead of `update` so that ActiveRecord callbacks are skipped. Some of the
callbacks are not desired here, for example, we do not want to create an audit
trail entry.

The nested AccessAndOpeningtimes class allows us to use the ActiveRecord query
interface despite the fact that the model has been removed from the app.