Use access and opening times columns
We have added columns to `worldwide_organisations` and `worldwide_offices` so
that we can remove the `access_and_opening_times` table to make it easier to
move the rendering of Worldwide Organisations and Worldwide Offices out of
Whitehall.

This replaces the use of the AccessAndOpeningTimes model with the new
attributes of the WorldwideOrganisations and WorldwideOffices. There is no
difference in functionality with this change.