Update national applicability tests for GOV.UK Design System
This updates the tests which ensure the national applicability
checkboxes work as intended.

This whole thing needs a rethink and refactor, but since this is
unblocking 1.6 i think this will do for now. I'm adding a card
to the backlog to rethink this.

This also fixes a bug which allowed a user to uncheck all the checkboxes
and save the edition. It now explicitly sets all_nation_applicability
to false if no checkboxes are present (was nil previously which
circumvented the validation.)