Enable Release 1.6 for all users
Forces the 'Preview design system' flag 'on' for all users, regardless
of their permissions.

We're doing this to make Release 1.6 of the Design System Transition to
all users at once.

However: we're skipping the override when running in the "test" env to
avoid breaking tests. Tests already cover both 'sides' of this flag, so
we're not losing any test coverage by doing this.