Remove unused Cucumber steps
I used the following sequence of commands to find steps that are not
used by either the default profile or the preview_design_system profile:

    bundle exec cucumber --dry-run --format stepdefs \
      | grep -B1 'NOT MATCHED' --no-group-separator \
      | grep features/step_definitions \
      > unused-steps-for-default-profile.txt

    bundle exec cucumber --dry-run --profile preview_design_system --format stepdefs \
      | grep -B1 'NOT MATCHED' --no-group-separator \
      | grep features/step_definitions \
      > unused-steps-for-preview-design-system-profile.txt

    comm -12 unused-steps-for-default-profile.txt \
      unused-steps-for-preview-design-system-profile.txt > unused-steps.txt

I've also removed some step helper methods which are no longer used.

Ideally I'd like to add check to the build to fail when there are any
unused steps, but I'll do that separately.