Remove unused abbreviation extractor script
This removes a script which was added to Whitehall in 2015, and its use
seems to have been lost in the mists of time.

The only documented use of `script/export_abbreviation_entities` is in
the now-archived repository, where it was referred to in the PR
description: alphagov/entity-manager#1.

It therefore seems very safe to consider it redundant code that can be
removed.