Remove an old single-use data migration script
The script `script/fix-content-id-replication` looks like it was only
ever required for a one-off data migration.

It was introduced in 2015 and hasn't changed since (except for updates
to keep in line with linter rules). And it's not referenced by any
other code repositories in the organisation. [1]

So I'm removing it as I'm quite confidence it's no longer in use.

[1]: https://github.com/search?q=org%3Aalphagov+fix-content-id-replication&type=code