Extract methods into Admin::AttachableHelper
I'm about to move the content & attachments for a WorldwideOrganisation
from the CorporateInformationPage about page onto the
WorldwideOrganisation itself. Once that move is completed, I will want
to add an Attachments tab and nav item to the WorldwideOrganisation
admin pages. In order to keep the UI consistent it makes sense to be
able to re-use the logic which adds the attachment count to the
Attachments tab label. This reduces some of the existing duplication and
it should be easier to add the tab and nav item to the
WorldwideOrganisation admin pages.