Ensure first_published_at param is cleaned on update
At the moment, there's an issue with the first_published_at value
not being set to nil when a user chooses 'This document has never been
published before.' for the document status radios.

When you update the radio from previously published to not previously
published the first_published_at params are still submitted on save. As
these params are not cleaned before save they continue to be persisted
on the edition.