Remove unused localisations of `find_an_embassy_title`
This localisation key is now only being used in the
EmbassiesIndexPresenter, for which we provide no means of selecting a
locale to use when generating the presented embassies_index
document[1]. This means these translations can be removed.

[1] In fact, we don't think these translations were ever being used
since the route that rendered the view that relied on them (removed in
the previous commit) didn't accept a `locale` parameter, so there was
no way a user could request that the page was rendered in a different
locale.