Add reorder_cabinet_ministers_roles endpoint
This is the first of 4 endpoints that will be used to reorder the
4 reorderable lists that used to live on the CabinetMinisters#show
page.

These pages will cover reordering:

- Cabinet minister ordering
- Ministers who also attend ordering
- Whips ordering
- Organisation ordering

This commit adds the Cabinet minister ordering page.

The reorderable list component passes params to the controller so I've
had rework the update_ordering method in the controller to accommodate
it.

I've also added some feature specs with reusable steps.