Remove Central African Republic embassy special case
https://trello.com/c/NTdQN6o3

FCDO have confirmed that we should direct people to the British Embassy
Kinshasa[1] on /world/embassies because this matches the foreign travel
advice[2]:

> Consular support is severely limited in CAR as there is no British
Embassy. However, the British Embassy Kinshasa, in the Democratic
Republic of Congo can provide limited remote consular support to British
nationals.

The data migration in this commit associates British Embassy Kinshasa
with the Central African Republic and then asserts that this is the only
embassy organisation association with this world location.

Note the test change that's necessary because it was previously relying
on the data in the Central African Republic special case.

[1]: https://www.gov.uk/world/organisations/british-embassy-kinshasa
[2]: https://www.gov.uk/foreign-travel-advice/central-african-republic