Remove French Polynesia embassy special case
https://trello.com/c/NTdQN6o3

FCDO have confirmed that we should direct people to the British Embassy
Paris[1] on /world/embassies because this matches the foreign travel
advice[2]:

> French Polynesia is a French Overseas Territory. There's no
British Embassy in French Polynesia. Consular support is limited.
However, the British Embassy in Paris, France can provide some consular
support to British nationals.

The data migration associates British Embassy Paris with French
Polynesia and then asserts that this is the only embassy organisation
associated with this World Location.

[1]: https://www.gov.uk/world/organisations/british-embassy-paris
[2]: https://www.gov.uk/foreign-travel-advice/french-polynesia