Remove Liechtenstein embassy special case
https://trello.com/c/NTdQN6o3

FCDO have confirmed that we should direct people to the British Embassy
Berne[1] on /world/embassies because this matches the foreign travel
advice[2]:

> Consular support is not available from the British government in
Liechtenstein. However the British Embassy in Berne, Switzerland can
provide consular support to British nationals.

FCDO have confirmed that there is no British Embassy in Liechtenstein so
we remove Liechtenstein from the British Embassy Liechtenstein[3] and
add British Embassy Berne to Liechtenstein before asserting that this is
the only embassy organisation associated with this location.

[1]: https://www.gov.uk/world/organisations/british-embassy-berne
[2]: https://www.gov.uk/foreign-travel-advice/liechtenstein
[3]: https://www.gov.uk/world/organisations/british-embassy-liechtenstein