Remove Macao embassy special case
https://trello.com/c/NTdQN6o3

FCDO have confirmed that we should direct people to the British
Consulate General Hong Kong[1] on /world/embassies because this matches the
foreign travel advice[2]:

> Consular support may be limited in Macao. However, the British
Consulate-General in Hong Kong can provide consular support to British
nationals.

FCDO have confirmed that there is no British Embassy in Macao so we
remove the Macao from the British Embassy Macao[3] and add British
Consulate General Hong Kong to Macao before asserting that this is the
only embassy associated with this location.

[1]: https://www.gov.uk/world/organisations/british-consulate-general-hong-kong
[2]: https://www.gov.uk/foreign-travel-advice/macao
[3]: https://www.gov.uk/world/organisations/british-embassy-macao