Remove Marshall Islands embassy special case
https://trello.com/c/NTdQN6o3

FCDO have confirmed that we should direct people to the British High
Commission Suva[1] on /world/embassies because this matches the foreign
travel advice[2]:

> There's no British Embassy in the Marshall Islands. Consular
support is limited, however the British High Commission Suva in Fiji can
provide some consular support to British nationals.

FCDO have confirmed that there is no British Embassy in the Marshall
Islands so we remove Marshall Islands from the British Embassy Marshall
Islands[3] and add British High Commission Suva to Marshall Islands
before asserting that this is the only embassy organisation associated
with this location.

[1]: https://www.gov.uk/world/organisations/british-high-commission-suva
[2]: https://www.gov.uk/foreign-travel-advice/marshall-islands
[3]: https://www.gov.uk/world/organisations/british-embassy-marshall-islands