Remove Micronesia embassy special case
https://trello.com/c/NTdQN6o3

FCDO have confirmed that we should direct people to the British High
Commission Suva[1] on /world/embassies because this matches the foreign
travel advice[2]:

> There's no British Embassy in Micronesia. Consular support is
severely limited, however, the British High Commission Suva in Fiji can
provide some consular support to British nationals.

FCDO have confirmed that there is no British Embassy in Micronesia so we
remove Micronesia from the British Embassy Micronesia[3]. We add British
High Commission Suva to Micronesia if neccesary and then assert that
this is the only embassy organisation associated with this location.

[1]: https://www.gov.uk/world/organisations/british-high-commission-suva
[2]: https://www.gov.uk/foreign-travel-advice/micronesia
[3]: https://www.gov.uk/world/organisations/british-embassy-micronesia