Remove New Caledonia embassy special case
https://trello.com/c/NTdQN6o3

FCDO have confirmed that we should direct people to the British Embassy
Paris[1] on /world/embassies because this matches the foreign
travel advice[2]:

> There's no British Embassy in New Caledonia. Consular support is
limited however, the British Embassy in Paris, France can provide
consular support to British nationals.

We add British Embassy Paris to New Caledonia if neccesary and then
assert that this is the only embassy organisation associated with this
location.

[1]: https://www.gov.uk/world/organisations/british-embassy-paris
[2]: https://www.gov.uk/foreign-travel-advice/new-caledonia