Remove St Martin embassy special case
https://trello.com/c/NTdQN6o3

FCDO have confirmed that we should direct people to the British Embassy
Paris[1] on /world/embassies because this matches the foreign travel
advice[2]:

> St Martin and St Barthélemy are French overseas territories. St
Martin is an Outermost Region of the European Union. Consular support is
not available from the British government in St Martin or St
Barthélemy. However, the British Embassy in Paris, France can provide
consular support to British nationals.

We add British Embassy Paris to St Martin if necessary and then assert
that this is the only embassy organisation associated with this
location.

[1]: https://www.gov.uk/world/organisations/british-embassy-paris
[2]: https://www.gov.uk/foreign-travel-advice/st-martin-and-st-barthelemy