Allow html to be passed into select-with-search component
Without this option text is presented as textContent so we see
sanitised output.

This has caused some bugs on various pages where organisations are
passed in some of which include '&' which is presented as &amp;.

This module only consumes options generated by Rails which have
been sanitised by default which rules out xss vulnerabilities.