Make Worldwide Office publishable
Worldwide Offices have been updated to have a `content_id` and Publishing API
presenter.

This includes the PublishesToPublishingApi module, so that Worldwide Offices
are published when changed.

The WorldwideOffice model currently delegates a bunch of methods to it's
related Contact. We need to add `content_id` to the list of excluded methods so
that the presenter is calling the method on the correct model.