Republish related Offices when Worldwide Organisations are changed
We want to make sure Worldwide Offices are republished when their parent
Worldwide Organisation changes. This will ensure any changes to default access
and opening times will be reflected in the offices that use them.

Because the new callback makes a call to `offices`,  we are now retrieving the
cached association in many of the tests. To fix this, we need to reload the
test objects.